Move 'NotificationDeleteConfirmation' to new package
