+++ app/k9mail/src/main/java/com/fsck/k9/notification/K9NotificationActionCreator.java
@@ -13,7 +13,7 @@ import com.fsck.k9.K9;
 import com.fsck.k9.R;
 import com.fsck.k9.activity.MessageList;
 import com.fsck.k9.controller.MessageReference;
-import com.fsck.k9.activity.NotificationDeleteConfirmation;
+import com.fsck.k9.ui.notification.NotificationDeleteConfirmation;
 import com.fsck.k9.activity.compose.MessageActions;
 import com.fsck.k9.activity.setup.AccountSetupIncoming;
 import com.fsck.k9.activity.setup.AccountSetupOutgoing;
